One new feature and two bug fixes.
Fixed name field text shown when a tree has nothing selected but has a
'none' string defined via the :include_blank option. Previously the given
text would only be shown if the user explicitly selected the tree's "none"
entry but not if the tree was shown for a resource where "none" applied
already (it always just used a blank string in this initial startup case).

Fixed silly mistake causing a failure to update the name field text of an
item selected in the root of the tree. The root nodes were being added
*before* the event handlers for changes of highlight were installed.

By request, added new :multi_expand option. Fully backwards compatible
with existing code. Normally a multiple selection tree allows multiple
nodes (branches) to open concurrently. A single selection tree only allows
one node (branch) to be open at a time. The multi_expand property can be
set to true or false to override this behaviour. The most obvious use is
to allow multiple branches of a single-selection tree to be concurrently
open. Setting the :multiple and :multi_expand options to the same value or
omitting the :multi_expand option altogether produces the same behaviour
as in previous versions of the plug-in.

Documentation is updated for a v1.10 release and an 'update.rb' script
has been added to make ugprades a little easier for existing users.